// source/components/plate-recognition-service/types.ts
export enum PrsEventType {
  REQUEST_IMAGE_ANALYSIS = "REQUEST_IMAGE_ANALYSIS",
  SUCCESSFUL_IMAGE_ANALYSIS = "SUCCESSFUL_IMAGE_ANALYSIS",
  UNSUCCESSFUL_IMAGE_ANALYSIS = "UNSUCCESSFUL_IMAGE_ANALYSIS",
}

export enum PrsPort {
  FROM_SEC = "FROM_SEC",
  TO_SEC = "TO_SEC",
  FROM_SXC = "FROM_SXC",
  TO_SXC = "TO_SXC",
}